Made it possible to set/get settings.
With this commit you can store/retrieve settings in the application.
There are two methods: Site::getSetting() & Site::setSetting(). When using
getSetting() you can specify a default returnvalue, which is returned when the
settings isn't found.

We use array-caching for the settings so there is a minimal amount of queries
to retrieve a single setting.