Convert a delete link into a form which will submit through POST
As we don't like CSRF we had to think about a secure way of deleting items, without being able to deleting an item just by accessing an URL.

Therefor I created this nifty piece of JS-code which will dynamically build a
form and will submit it after the confirmation is shown (and confirmed).

All you need to do is add a class `confirmDelete` on the link. The
href-attribute will be used as the action for the form.

If you need to pass extra fields (which will be the case most of the time) you
can add extra (hidden-)fields by passing them as data-attributes. The
data-attributes that start with `field` will be converted into hidden fields.

For example: `data-field-form="delete"` will become
`<input type="hidden" name="form" value="delete" />`

So, from now on delete-actions should be an action that acts like it is a form.

A very basic example of such an action could be

	$frm = new SiteForm('delete', null, 'post', true);
	$frm->addHidden('id');

	if($frm->isSubmitted()) {
		$frm->getField('id')->isFilled('invalid');
		if($frm->isCorrect()) {
			$id = $frm->getField('id')->getValue();
			$item = User::get($id);
			if(!$item) Site::displayError('Not found', 404);
			$item->delete();

			$this->redirect(
				$this->url->buildUrl(
					'index', null, null,
					array('report' => 'deleted', 'var' => $item->getName())
				)
			);
		}
		else Site::displayError('Bad request', 400);
	}
	else Site::displayError('Bad request', 400);

And a delete link will look like:

	<a href="{$var|buildurl:'delete':'<module>'}"
		class="btn btn-danger confirmPostForm"
		data-field-form="delete" data-field-id="{$item.id}"
		data-message="{$msgAreYouSure}"
	>
		<i class="icon-trash icon-white"></i>
		{$lblDelete|ucfirst}
	</a>