Automatic handling for sequence by drag a drop.
It works the same as in Fork, as in:

You can pass the action, module through data-attributes on the datagrid. By
default we will use "sequence" and the current module.

When done dragging an AJAX-request will be made to
/ajax.php?module=[module]&action=[action]&language=[current_language].

If your code returns 200 a decent message will be shown, if not an error
will be shown and the drag and drop will be reverted.